feat: upgrade to quince
