--- changelog.d/20231205_113430_regis_local_edly_io.md
@@ -1,1 +1,0 @@
-- 💥[Feature] Replace "*.local.overhang.io" hostnames by "*.local.edly.io". (by @regisb)
