+++ requirements/base.txt
@@ -6,27 +6,27 @@
 #
 appdirs==1.4.4
     # via -r requirements/base.in
-cachetools==5.3.1
+cachetools==5.3.2
     # via google-auth
 certifi==2023.7.22
     # via
     #   kubernetes
     #   requests
-charset-normalizer==3.2.0
+charset-normalizer==3.3.1
     # via requests
 click==8.1.7
     # via -r requirements/base.in
-google-auth==2.22.0
+google-auth==2.23.3
     # via kubernetes
 idna==3.4
     # via requests
 jinja2==3.1.2
     # via -r requirements/base.in
-kubernetes==27.2.0
+kubernetes==28.1.0
     # via -r requirements/base.in
 markupsafe==2.1.3
     # via jinja2
-mypy==1.5.1
+mypy==1.6.1
     # via -r requirements/base.in
 mypy-extensions==1.0.0
     # via mypy
@@ -40,7 +40,7 @@ pyasn1==0.5.0
     #   rsa
 pyasn1-modules==0.3.0
     # via google-auth
-pycryptodome==3.18.0
+pycryptodome==3.19.0
     # via -r requirements/base.in
 python-dateutil==2.8.2
     # via kubernetes
@@ -58,19 +58,17 @@ rsa==4.9
     # via google-auth
 six==1.16.0
     # via
-    #   google-auth
     #   kubernetes
     #   python-dateutil
 tomli==2.0.1
     # via mypy
-typing-extensions==4.7.1
+typing-extensions==4.8.0
     # via
     #   -r requirements/base.in
     #   mypy
 urllib3==1.26.18
     # via
-    #   google-auth
     #   kubernetes
     #   requests
-websocket-client==1.6.2
+websocket-client==1.6.4
     # via kubernetes
