+++ requirements/plugins.txt
@@ -1,12 +1,13 @@
-# change version ranges when upgrading from palm
-tutor-android>=16.0.0,<17.0.0
-tutor-cairn>=16.0.0,<17.0.0
-tutor-discovery>=16.0.0,<17.0.0
-tutor-ecommerce>=16.0.0,<17.0.0
-tutor-forum>=16.0.0,<17.0.0
-tutor-license>=16.0.0,<17.0.0
-tutor-mfe>=16.0.0,<17.0.0
-tutor-minio>=16.0.0,<17.0.0
-tutor-notes>=16.0.0,<17.0.0
-tutor-webui>=16.0.0,<17.0.0
-tutor-xqueue>=16.0.0,<17.0.0
+# change version ranges when upgrading from quince
+tutor-android>=17.0.0,<18.0.0
+tutor-cairn>=17.0.0,<18.0.0
+tutor-credentials>=17.0.0,<18.0.0
+tutor-discovery>=17.0.0,<18.0.0
+tutor-ecommerce>=17.0.0,<18.0.0
+tutor-forum>=17.0.0,<18.0.0
+tutor-jupyter>=17.0.0,<18.0.0
+tutor-mfe>=17.0.0,<18.0.0
+tutor-minio>=17.0.0,<18.0.0
+tutor-notes>=17.0.0,<18.0.0
+tutor-webui>=17.0.0,<18.0.0
+tutor-xqueue>=17.0.0,<18.0.0
