+++ tutor/commands/upgrade/__init__.py
@@ -1,4 +1,4 @@
-# Note: don't forget to change this when we upgrade from palm
+# Note: don't forget to change this when we upgrade from quince
 OPENEDX_RELEASE_NAMES = [
     "ironwood",
     "juniper",
@@ -7,4 +7,6 @@ OPENEDX_RELEASE_NAMES = [
     "maple",
     "nutmeg",
     "olive",
+    "palm",
+    "quince",
 ]
