+++ tutor/templates/dev/docker-compose.yml
@@ -12,8 +12,6 @@ x-openedx-service:
     - ../apps/openedx/config:/openedx/config:ro
     # theme files
     - ../build/openedx/themes:/openedx/themes
-    # editable requirements
-    - ../build/openedx/requirements:/openedx/requirements
 
 services:
   permissions:
